重新排版了代码。接下来查 * thread #1, queue = 'com.apple.main-thread', stop reason = EXC_BAD_ACCESS (code=2, address=0x14fdfffe0)     frame #0: 0x00000001a132b5a0 libsystem_malloc.dylib`tiny_malloc_should_clear + 8 libsystem_malloc.dylib`tiny_malloc_should_clear: ->  0x1a132b5a0 <+8>:  stp    x28, x27, [sp, #0x90]     0x1a132b5a4 <+12>: stp    x26, x25, [sp, #0xa0]     0x1a132b5a8 <+16>: stp    x24, x23, [sp, #0xb0]     0x1a132b5ac <+20>: stp    x22, x21, [sp, #0xc0] 的BUG

+++ run.sh
@@ -1,12 +1,25 @@
 #!/bin/zsh
-rm -rf ./executable
-rm -rf ./output
+
+# 运行脚本，用于批量测试CNF算例
+
+if test -e "./executable"
+then
+    rm -rf ./executable
+fi
+
+if test -e "./output" 
+then
+    rm -rf ./output
+fi
+
 mkdir executable
 mkdir output
-read
+
 g++ ./src/main.cpp -o ./executable/dpll -std=c++11 -O0 -Wl,-stack_size -Wl,0x20000000 
 g++ ./src/verify.cpp -o ./executable/verify -std=c++11
+
 echo "\n\nSnapshot:"
+
 for file in `ls  ./testset` 
     do
     echo "\nSolving ${file}"
